feat: 16-bit output; fix a red/blue swap in the screenshot harness
The graph ended in RGBA8Unorm, so every export was eight bits per channel
whatever container it went into -- a gradient that survived the whole pipeline
in float got quantised at the very last node. develop:display and geometry now
write RGBA16Float, and export converts to 16-bit unsigned for the encoder. Half
float rather than 16-bit integer because Metal guarantees it read-write, and
because it holds the shadows better than a uniform integer scale.

Everything that read the output had to move with it: the colour picker, the
histogram, export, the export-size measurement, the bench, and the screenshot
harness.

Fixing the harness turned up a bug in it. Its CGImage used byteOrder32Little
with premultipliedFirst, which reads the pipeline's RGBA bytes as BGRA -- so
every screenshot taken tonight had red and blue swapped. The test frame is a
night sky, and it was rendering warm brown. It is blue now, which is what a
night sky is. The app itself was never affected; its blit samples the texture
through Metal rather than reinterpreting bytes.

The new test renders a gradient fine enough that eight bits could not represent
it and counts the distinct values that come back. Changing a texture format is
the kind of edit that looks done the moment it compiles.

Tests: 110 engine checks, 1933 viewport.

+++ engine/src/util/ImageWriter.h
@@ -20,7 +20,7 @@ struct ExportOptions {
 /// Resampling, when requested, happens in CoreGraphics rather than on the GPU:
 /// export is not on the interaction path, and a correct downscale with proper
 /// filtering matters more here than shaving milliseconds.
-void writeImage(const std::string& path, const std::uint8_t* rgba,
+void writeImage(const std::string& path, const std::uint16_t* rgba,
                 std::uint32_t width, std::uint32_t height, std::size_t bytesPerRow,
                 const ExportOptions& options);
 
@@ -28,12 +28,12 @@ void writeImage(const std::string& path, const std::uint8_t* rgba,
 /// Encodes to memory and reports the byte count, without writing anything.
 /// The export panel needs a size it can trust before you commit to the write,
 /// and only a real encode gives one.
-[[nodiscard]] std::size_t encodedSize(const std::uint8_t* rgba,
+[[nodiscard]] std::size_t encodedSize(const std::uint16_t* rgba,
                                       std::uint32_t width, std::uint32_t height,
                                       std::size_t bytesPerRow,
                                       const ExportOptions& options);
 
-void writePng(const std::string& path, const std::uint8_t* rgba,
+void writePng(const std::string& path, const std::uint16_t* rgba,
               std::uint32_t width, std::uint32_t height, std::size_t bytesPerRow);
 
 /// Picks a format from the path's extension, defaulting to JPEG.
